backend/requirements.txt
psycopg2-binary
django
djangorestframework
django-enumchoicefield
celery
redis
requests
pycountry